🧬 Migrate the tests about tesing generating Python code as another testing module.
