+++ test/unit_test/server/application/init.py
@@ -72,14 +72,14 @@ class AppServerTestSpec(metaclass=ABCMeta):
                 web_app, self.expected_sut_type
             ), f"The web application server it generates should be *{self.expected_sut_type}* type object."
 
-    def test_generate_pycode_about_annotating_function(self, sut: BaseAppServer):
-        for_test_api_name = "/Function/name"
-        api = Mock(MockAPI(url=Mock(), http=Mock()))
-        api.http.request.parameters = [APIParameter().deserialize(p) for p in _Test_API_Parameters]
-
-        annotate_function_pycode = sut._annotate_function(api_name=for_test_api_name, api_config=api)
-
-        assert sut._api_controller_name(for_test_api_name) in annotate_function_pycode
+    # def test_generate_pycode_about_annotating_function(self, sut: BaseAppServer):
+    #     for_test_api_name = "/Function/name"
+    #     api = Mock(MockAPI(url=Mock(), http=Mock()))
+    #     api.http.request.parameters = [APIParameter().deserialize(p) for p in _Test_API_Parameters]
+    #
+    #     annotate_function_pycode = sut._annotate_function(api_name=for_test_api_name, api_config=api)
+    #
+    #     assert sut._api_controller_name(for_test_api_name) in annotate_function_pycode
 
     @pytest.mark.parametrize(
         ("method", "api_params", "error_msg_like", "expected_status_code"),
@@ -139,42 +139,42 @@ class AppServerTestSpec(metaclass=ABCMeta):
     def _expected_response_type(self) -> Type[Any]:
         pass
 
-    @pytest.mark.parametrize("base_url", [None, "Has base URL"])
-    def test_generate_pycode_about_adding_api(self, sut: BaseAppServer, base_url: Optional[str]):
-        # TODO: Should think a better implementation of this test
-        for_test_api_name = "Function name"
-        for_test_url = "this is an url path"
-        for_test_req_method = "HTTP method"
-        api_config = Mock(MockAPI(url=Mock(), http=Mock()))
-        api_config.url = for_test_url
-        api_config.http.request.method = for_test_req_method
+    # @pytest.mark.parametrize("base_url", [None, "Has base URL"])
+    # def test_generate_pycode_about_adding_api(self, sut: BaseAppServer, base_url: Optional[str]):
+    #     # TODO: Should think a better implementation of this test
+    #     for_test_api_name = "Function name"
+    #     for_test_url = "this is an url path"
+    #     for_test_req_method = "HTTP method"
+    #     api_config = Mock(MockAPI(url=Mock(), http=Mock()))
+    #     api_config.url = for_test_url
+    #     api_config.http.request.method = for_test_req_method
+    #
+    #     add_api_pycode = sut._add_api(
+    #         api_name=for_test_api_name, api_config=self._get_api_config_param(api_config), base_url=base_url
+    #     )
+    #
+    #     assert (
+    #         self._get_url_criteria(base_url) in add_api_pycode
+    #         and self._get_http_method_in_generating_code(for_test_req_method) in add_api_pycode
+    #     )
+    #
+    # @abstractmethod
+    # def _get_api_config_param(self, api_config: MockAPI) -> Union[MockAPI, List[MockAPI]]:
+    #     pass
+    #
+    # @abstractmethod
+    # def _get_url_criteria(self, base_url: Optional[str]) -> str:
+    #     pass
+    #
+    # @abstractmethod
+    # def _get_http_method_in_generating_code(self, method: str) -> str:
+    #     pass
 
-        add_api_pycode = sut._add_api(
-            api_name=for_test_api_name, api_config=self._get_api_config_param(api_config), base_url=base_url
-        )
-
-        assert (
-            self._get_url_criteria(base_url) in add_api_pycode
-            and self._get_http_method_in_generating_code(for_test_req_method) in add_api_pycode
-        )
-
-    @abstractmethod
-    def _get_api_config_param(self, api_config: MockAPI) -> Union[MockAPI, List[MockAPI]]:
-        pass
-
-    @abstractmethod
-    def _get_url_criteria(self, base_url: Optional[str]) -> str:
-        pass
-
-    @abstractmethod
-    def _get_http_method_in_generating_code(self, method: str) -> str:
-        pass
-
-    def test__record_api_params_info_with_invalid_value(self, sut: BaseAppServer):
-        ut_url = "This is URL path"
-        ut_api_config = "Invalid API configuration"
-        with pytest.raises(TypeError):
-            sut._record_api_params_info(url=ut_url, api_config=ut_api_config)
+    # def test__record_api_params_info_with_invalid_value(self, sut: BaseAppServer):
+    #     ut_url = "This is URL path"
+    #     ut_api_config = "Invalid API configuration"
+    #     with pytest.raises(TypeError):
+    #         sut._record_api_params_info(url=ut_url, api_config=ut_api_config)
 
 
 class TestFlaskServer(AppServerTestSpec):
@@ -207,21 +207,21 @@ class TestFlaskServer(AppServerTestSpec):
     def _expected_response_type(self) -> Type[FlaskResponse]:
         return FlaskResponse
 
-    def _get_api_config_param(self, api_config: MockAPI) -> List[MockAPI]:
-        return [api_config]
+    # def _get_api_config_param(self, api_config: MockAPI) -> List[MockAPI]:
+    #     return [api_config]
+    #
+    # def _get_url_criteria(self, base_url: Optional[str]) -> str:
+    #     return "Function name"
+    #
+    # def _get_http_method_in_generating_code(self, method: str) -> str:
+    #     return method
 
-    def _get_url_criteria(self, base_url: Optional[str]) -> str:
-        return "Function name"
-
-    def _get_http_method_in_generating_code(self, method: str) -> str:
-        return method
-
-    def test__add_api_with_invalid_value(self, sut: BaseAppServer):
-        ut_url = "This is URL path"
-        ut_api_config = MockAPI(url=ut_url)
-        with patch.object(sut, "_record_api_params_info"):
-            with pytest.raises(TypeError):
-                sut._add_api(api_name=ut_url, api_config=ut_api_config)
+    # def test__add_api_with_invalid_value(self, sut: BaseAppServer):
+    #     ut_url = "This is URL path"
+    #     ut_api_config = MockAPI(url=ut_url)
+    #     with patch.object(sut, "_record_api_params_info"):
+    #         with pytest.raises(TypeError):
+    #             sut._add_api(api_name=ut_url, api_config=ut_api_config)
 
 
 class TestFastAPIServer(AppServerTestSpec):
@@ -269,22 +269,22 @@ class TestFastAPIServer(AppServerTestSpec):
     def _expected_response_type(self) -> Type[FastAPIResponse]:
         return FastAPIResponse
 
-    def _get_api_config_param(self, api_config: MockAPI) -> MockAPI:
-        return api_config
+    # def _get_api_config_param(self, api_config: MockAPI) -> MockAPI:
+    #     return api_config
+    #
+    # def _get_url_criteria(self, base_url: Optional[str]) -> str:
+    #     for_test_url = "this is an url path"
+    #     return f"{base_url}{for_test_url}" if base_url else for_test_url
+    #
+    # def _get_http_method_in_generating_code(self, method: str) -> str:
+    #     return method.lower()
 
-    def _get_url_criteria(self, base_url: Optional[str]) -> str:
-        for_test_url = "this is an url path"
-        return f"{base_url}{for_test_url}" if base_url else for_test_url
-
-    def _get_http_method_in_generating_code(self, method: str) -> str:
-        return method.lower()
-
-    def test__add_api_with_invalid_value(self, sut: BaseAppServer):
-        ut_url = "This is URL path"
-        ut_api_config = ["Invalid API configuration"]
-        with patch.object(sut, "_record_api_params_info"):
-            with pytest.raises(TypeError):
-                sut._add_api(api_name=ut_url, api_config=ut_api_config)
+    # def test__add_api_with_invalid_value(self, sut: BaseAppServer):
+    #     ut_url = "This is URL path"
+    #     ut_api_config = ["Invalid API configuration"]
+    #     with patch.object(sut, "_record_api_params_info"):
+    #         with pytest.raises(TypeError):
+    #             sut._add_api(api_name=ut_url, api_config=ut_api_config)
 
 
 class TestInnerHTTPResponse:
